+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes in 4.9
 - Only display Categories having fields in AssignOtherInfo.php
 - Accessibility: add missing input label in AssignOtherInfo.php
 - PROGRAM_CONFIG table: Add Allow Teachers to edit gradebook grades for past quarters option in Update.fnc.php, rosariosis.sql & Configuration.php, sponsored by Santa Cecilia school
+- Accessibility: add missing input label in Widgets.fnc.php
 
 Changes in 4.8
 --------------